Made evaluate a common function as part of utils.py(new file) in common
