+++ jigglypuffRL/common/__init__.py
@@ -2,3 +2,4 @@ from jigglypuffRL.common.policies import BasePolicy, MlpPolicy, get_policy_from_
 from jigglypuffRL.common.values import MlpValue, get_value_from_name
 from jigglypuffRL.common.actor_critic import ActorCritic, MlpActor, MlpCritic
 from jigglypuffRL.common.buffers import ReplayBuffer
+from jigglypuffRL.common.utils import evaluate
